Move some types into `db/types`
