+++ backend/src/api/id.rs
@@ -1,49 +1,8 @@
-use bytes::BytesMut;
 use paste::paste;
-use postgres_types::{FromSql, ToSql};
 use static_assertions::const_assert;
 use std::{convert::TryInto, fmt};
 
-
-/// The type of key we are using. Implements `ToSql` and `FromSql` by casting
-/// to/from `i64`.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
-pub(crate) struct Key(pub(crate) u64);
-
-impl ToSql for Key {
-    fn to_sql(
-        &self,
-        ty: &postgres_types::Type,
-        out: &mut BytesMut,
-    ) -> Result<postgres_types::IsNull, Box<dyn std::error::Error + Sync + Send>> {
-        (self.0 as i64).to_sql(ty, out)
-    }
-
-    fn accepts(ty: &postgres_types::Type) -> bool {
-        <i64 as ToSql>::accepts(ty)
-    }
-
-    fn to_sql_checked(
-        &self,
-        ty: &postgres_types::Type,
-        out: &mut postgres_types::private::BytesMut,
-    ) -> Result<postgres_types::IsNull, Box<dyn std::error::Error + Sync + Send>> {
-        (self.0 as i64).to_sql_checked(ty, out)
-    }
-}
-
-impl<'a> FromSql<'a> for Key {
-    fn from_sql(
-        ty: &postgres_types::Type,
-        raw: &'a [u8],
-    ) -> Result<Self, Box<dyn std::error::Error + Sync + Send>> {
-        i64::from_sql(ty, raw).map(|i| Key(i as u64))
-    }
-
-    fn accepts(ty: &postgres_types::Type) -> bool {
-        <i64 as FromSql>::accepts(ty)
-    }
-}
+use crate::db::types::Key;
 
 
 /// An opaque, globally-unique identifier for all "nodes" that the GraphQL API
