+++ backend/src/api/model/event.rs
@@ -3,7 +3,10 @@ use futures::stream::TryStreamExt;
 use tokio_postgres::Row;
 use juniper::{GraphQLObject, graphql_object, FieldResult};
 
-use crate::api::{Context, Id, db::EventTrack, id::Key, model::series::Series};
+use crate::{
+    api::{Context, Id, model::series::Series},
+    db::types::{EventTrack, Key},
+};
 
 
 pub(crate) struct Event {
